klient/mount: move index package from mount to machine folder
