+++ go/src/koding/klient/machine/index/cache_test.go
