+++ go/src/koding/klient/machine/index/index_test.go
