+++ go/src/koding/klient/machine/index/index_unix_test.go
