+++ go/src/koding/klient/machine/index/index_windows_test.go
