+++ go/src/koding/klient/machine/index/kite.go
